Add new selector for div.text.child
I wanted to discover a way to make this work without
having to write a new selector, because the instructions
imply that that is possible, and because it is a greater
challenge, it seems more desirable from a problem-solving
standpoint.

But I also don't want to get lost in the weeds here. I need
to keep moving forward and I can always revisit this
exercise later.